scripts/photo-intelligence/requirements.txt
# Core ML and image processing
face-recognition>=1.3.0
mediapipe>=0.10.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Deep learning models
ultralytics>=8.0.0  # YOLOv8 for person detection
sentence-transformers>=2.2.0  # CLIP embeddings
torch>=2.0.0
torchvision>=0.15.0

# Perceptual hashing
imagehash>=4.3.0

# Clustering
hdbscan>=0.8.33
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0




